requirements.txt
pandas==2.2.2
numpy==1.26.4
ydata-profiling==4.8.1
python-dotenv==1.0.1
pytest==8.2.0
pyarrow
pandera==0.19.0